Added multi-architecture test cases
The test case build system now uses gasm to support
the following target platforms:

- aarch64
- armv5T (I think...)
- armv7
- mips32r2 (WLOG: big-endian)
- mips64r2 (WLOG: big-endian)
- powerpc
- powerpc64
- riscv64
- sparc64

The makefile will build for all target platforms by default.
It will test to see if the appropriate tools are available,
and produce an error if they are not.

To bypass this feature and only build x86-64 tests with nasm,
run `make nasm`

There are now cross-arch examples of the following tests:

- recursion
- square
- stack
- strlen