Unified multi-arch support for Small World
This shifts away from the attempt to auto-generate state,
since the existing models are rarely sufficient to do that.

This adds a number of new CPU models,
as well as rearchitecting the Unicorn emulator
to capture all parameters in a data class.

This is tested on amd64, armv5t, armv7a, aarch64, and mips32.
It's also tested on mips64, but that explodes violently.