Fixed some regressions with heap allocation
Concretizing values in BumpAllocator.malloc()
disabled a nice feature where you could modify the heap
by modifying a ctype Struct stored on the heap.
That feature is now reenabled; need to ask if
_not_ immediately concretizing the heap causes issues.

Bug in BumpAllocator.value meant that if the heap wasn't full,
the value would be shorter than the total size.
This could make loading the heap back from Unicorn fail
if value occupied fewer pages than the entire heap.
Previous tests used a heap that was too small to catch this.