Resolved problem with colorizer
The instability is a known problem.
Identifying congruent behaviors across non-deterministic runs
is a hard problem, for which we don't have a solution.

This commit forces colorizer runs to execute deterministically
by seeding the PRNG that generates the colors.

This weakens the colorizer's ability to explore program state,
but we're working on that.