ELF parser can now determine platform
This uses the ELF machine type and other header fields
to determine architecture and byteorder

This can be compared against a desired platform
for verification, or used to determine
a platform for execution.

It's not 100% accurate;
ARM32 is vague about ABI versions after 5 and its CPU series.
MIPS is accurate, but won't detect rev6 or other oddities