Altered how unicorn 'returns' from function hooks.
Previously, it worked by tracking the current PC,
and inserting execution at the next instruction
once the code callback was complete.

This made the big assumption that the function
was invoked expecting a standard, single-frame call and return.
There are cases where the invoked function
returns to a previous frame; possibly several frames in the past.

I fixed this issue by instead having `code_callback()`
emulate an architecture-specific return operation.