Numerous fixes to get powerpc64 working
PowerPC64 doesn't use normal function pointers.
Instead, it points to a "function descriptor" struct,
which contains the address and a "table of contents" pointer
for the containing ELF file.

This means that the entrypoint address
in the ELF header doesn't actually point to code,
and that the relocation processor is only partially correct;
the JUMP_SLOT relocation type fills in the entire 24-byte
function descriptor.

Also in this commit, I changed ElfExecutable
to use a RangeCollection for its bounds.