Added a sparse IO stream class
AngrEmulator needs to convert Executable objects into an IO stream
to interface with CLE's program loader.

The only Python class for doing this is BytesIO,
which provides a stream interface to a single, non-sparse
bytes-like object.

This is fine for small programs, but causes huge problems
if you try to load a RAM image.