Extended smallworld API to accept symbolic bitvector expressions.
Almost all original usage patterns are unchanged;
the only exception was a change to the signature
of memory read callbacks.

The State API is almost entirely unchanged,
and all original usage should be valid.
The one exception is an incidental change to the signature
of memory read hooks.

The Emulator API handles things as follows:
- Writing content can take a concrete or symbolic value
- Writing symbolic content to an unsupported emulator raises an exception.
- The original content-read APIs only return concrete values.
- If the value is symbolic and can't be collapsed, it raises an exception.
- There's a second content-read API for reading symbolic values.
- There are two memory hook APIs, for reading concrete and symbolic values.