feat: Added basic ELF library linking (#175)
This adds a utility function to the ELFExecutable class that allows it to update all undefined symbols using the values of defined symbols from a second ELFExecutable object. This mimics the static/dynamic linker.

NOTE: Currently, this won't handle all features of the dynamic linker, and won't handle especially-crazy relocation types such as TLS entries.