Partial commit of a lot of bug and edge case fixes.
Exit points now must be inside mapped memory,
thanks to a change in Unicorn and Panda's handling.

This is going to require touching essentially every test script,
and almost all of the assembly test cases.
I got through all the easy ones.  The rest rely on symbol sizes,
which are coming next commit.